chore: update directory
